pyproject.toml
[project]
name = "flask-api-intro"
version = "0.1.0"
description = "Flask API - Introduction to Building APIs with Flask"
authors = [
    { name = "Baboyma Kagniniwa", email = "baboyma@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "flask>=3.1.1",
    "flask-restful>=0.3.10",
    "flask-sqlalchemy>=3.1.1",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
]

[project.scrits]
app = "flask_api_intro.app:main"
